Реализована функция encrypt_caesar()
