+++ pgocs102
@@ -1,0 +1,2 @@
+cd /c/Users/ksenk/Documents/cs102-2025-1
+
